Report mockups + delete button

+++ src/main/resources/static/app/file-demo/file-page.controller.js
@@ -3,7 +3,6 @@ angular.module('messageHubDemoApp')
         function ($scope, fileUpload) {
             //The container for watching the file uploads.
             $scope.fileUploads = {};
-
             //The value from the file input.
             $scope.myFile = {};
 
@@ -20,10 +19,70 @@ angular.module('messageHubDemoApp')
                 });
             }
 
-            $scope.removeFile = function(fileId) {
-                fileUpload.removeFile(fileId);
-                delete $scope.fileUploads[fileId];
+            $scope.cancelRemove = function(fileId) {
+                console.log("UNDO Removing file: " + fileId);
+                var timeout = $scope.fileUploads[fileId].deleteTimeout;
+                clearTimeout(timeout);
+                $scope.fileUploads[fileId].deleteRequested = false;
+                $scope.fileUploads[fileId].countdown = 0;
+
+                fileUpload.unpauseFileProcess(fileId);
+                $scope.fileUploads[fileId].status = 'started';
+                watchFileProcess(fileId);
+                //
+                //if ($scope.fileUploads[fileId].percentage < 100) {
+                //        fileUpload.reportFileProcess(fileId)
+                //            .success(function (data) {
+                //                console.log("file is finished? : " + !(data.percentage < 100));
+                //                console.log("file percent? : " + data.percentage);
+                //                if (data.percentage < 100) {
+                //                    //Once the data is returned, save it, and take a look.
+                //
+                //                } else {
+                //                    $scope.fileUploads[fileId] = data;
+                //                    $scope.fileUploads[fileId].status = "finished";
+                //                }
+                //            });
+                //} else {
+                //    $scope.fileUploads[fileId].status = "finished";
+                //}
+
             };
+
+            $scope.removeFile = function(fileId) {
+
+                if ($scope.fileUploads[fileId] && $scope.fileUploads[fileId].deleteRequested) {
+                    console.log("Remove already requested for: " + fileId);
+                    //it's already marked for being deleted.
+                    return;
+                }
+
+                $scope.fileUploads[fileId].deleteRequested = true;
+                $scope.fileUploads[fileId].status = 'deleting';
+                $scope.fileUploads[fileId].countdown = 100;
+                fileUpload.pauseFileProcess(fileId);
+                countdown(fileId);
+
+                $scope.fileUploads[fileId].deleteTimeout = setTimeout(function () {
+                    if ($scope.fileUploads[fileId].deleteRequested === true) {
+                        console.log("Removing file: " + fileId);
+                        delete $scope.fileUploads[fileId];
+                        fileUpload.removeFile(fileId);
+                    }
+                }, 2200);
+            };
+
+            function countdown(fileId) {
+                setTimeout(function() {
+                    if ($scope.fileUploads[fileId]) {
+                        $scope.fileUploads[fileId].countdown = Math.round($scope.fileUploads[fileId].countdown - 3.8);
+                        if ($scope.fileUploads[fileId].countdown > 0) {
+                            $scope.$apply();
+                            countdown(fileId);
+                        }
+                    }
+                }, 100);
+            }
 
             //The function that sends the file and watches the process.
             $scope.uploadFile = function () {
@@ -51,29 +110,37 @@ angular.module('messageHubDemoApp')
             };
 
             function watchFileProcess(fileId) {
+
                 setTimeout(function () {
-                    fileUpload.reportFileProcess(fileId)
-                    //Once the data is returned, save it, and take a look.
-                        .success(function (data) {
-                            if (data) {
-                                //replace the data in the map with the new information about the process.
-                                $scope.fileUploads[fileId] = data;
-                                //And if it's not done, trigger another update.
-                                if (data.status === 'started') {
-                                    //Call again to get an update.  this is recursion. w00t!
-                                    watchFileProcess(fileId);
+                    var upload = $scope.fileUploads[fileId];
+                    if (!upload || upload.status !== 'deleting') {
+                        //console.log("watching...");
+                        //console.log(upload);
+                        fileUpload.reportFileProcess(fileId)
+                        //Once the data is returned, save it, and take a look.
+                            .success(function (data) {
+                                if (data) {
+                                    //replace the data in the map with the new information about the process.
+                                    $scope.fileUploads[fileId] = data;
+                                    //And if it's not done, trigger another update.
+                                    if (data.status === 'started') {
+                                        //Call again to get an update.  this is recursion. w00t!
+                                        watchFileProcess(fileId);
+                                    }
                                 }
-                            }
-                        });
+                            });
+                    }
                     //repeat the lookup at the time interval below.
-                }, 1000);
-            }
+                }, 500);
+        }
 
             $scope.downloadAcks = function(fileId) {
-                var fileInfo = $scope.fileUploads[fileId];
-                var fileName = fileInfo.fileName;
-                var text = fileInfo.ackMessages.join("\r");
-                download(fileName, text);
+                fileUpload.getAcks(fileId)
+                    .success(function (data) {
+                        var fileName = $scope.fileUploads[fileId].fileName;
+                        var text = data.join("\r");
+                        download(fileName, text);
+                    });
             };
 
             function download(filename, text) {
